+++ src/main/java/com/lescastcodeurs/bot/StringUtils.java
@@ -9,6 +9,14 @@ public final class StringUtils {
     // prevent instantiation
   }
 
+  public static boolean isBlank(String s) {
+    return s == null || s.isBlank();
+  }
+
+  public static boolean isNotBlank(String s) {
+    return !isBlank(s);
+  }
+
   public static String normalize(String s) {
     String normalized = Normalizer.normalize(s, Normalizer.Form.NFD);
     normalized = normalized.replaceAll("\\p{InCombiningDiacriticalMarks}", "");