+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Improve `GENERATE_SHOW_NOTES` command help (#18).
 - Relax constraints on show notes messages (#20). All messages containing at least a link are now copied in the show
   notes.
+- Exclude Slack bot and application messages from show notes (#32).
 
 ### Fixed